refactor(landing): one file per blog post, auto-discovered
Move each post into its own self-contained file under src/blog/posts/<slug>.tsx
(filename = slug), exporting its own `meta` + default `Body`. `src/blog/index.ts`
collects them with `import.meta.glob` (sorted newest-first), and the prerender
config enumerates slugs via `fs.readdirSync` — so adding a post is dropping in
one file, with no central list, `registry.ts`, or `bodies` map to maintain.
